Add filter by status method
Created the method to filter the todos by status so you can view all the complete ones or all the incomplete ones. Had to change the tests to have the entry complete or incomplete since we originally thought it would be true or false until we checked the html file. Co-authored-by: Dante Miller mill8622@morris.umn.edu